% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testPlotly}
\alias{testPlotly}
\title{Test plotly object}
\usage{
testPlotly(app, id, data, interval = 1000, timeout = 30000)
}
\arguments{
\item{app}{shinytest/shinytest2 application object (ShinyDriver/AppDriver)}

\item{id}{string, id of the plotly object}

\item{data}{list/data.frame, reference data}

\item{interval}{see: \link[shinytest]{ShinyDriver} (method waitFor(), param checkInterval)
or \link[shinytest2]{AppDriver} (method wait_for_js(), param interval)}

\item{timeout}{see: \link[shinytest]{ShinyDriver} (method waitFor(), param timeout)
or \link[shinytest2]{AppDriver} (method wait_for_js(), param timeout)}
}
\value{
ToDo
}
\description{
Test plotly object
}
